care_v0.2.0_release/care/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "care"
version = "0.2.0"
description = "Curvature-Aware Risk Engine — attractor-based infrastructure hardening"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Zetta Byte" }]
keywords = ["security", "risk", "curvature", "attractor", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
jax = [
    "jax>=0.4.25",
    "jaxlib>=0.4.25",
]
hcderiv = [
    "hcderiv[jax]>=0.4.0",
]
viz = [
    "matplotlib>=3.8",
    "plotly>=5.20",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.9",
    "care[viz]",
]
all = [
    "care[jax,hcderiv,viz,dev]",
]

[project.scripts]
care-server = "care.api.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["care*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
